clean up testutils
