+++ testutils/package.json
@@ -11,7 +11,8 @@
   "files": [
     "lib/*.d.ts",
     "lib/*.js.map",
-    "lib/*.js"
+    "lib/*.js",
+    "default.ipynb"
   ],
   "main": "lib/index.js",
   "types": "lib/index.d.ts",
